chore: updated migration fields

+++ apps/backend/migrations/1769615639231_create_auth_tokens_table.ts
@@ -10,12 +10,12 @@ export async function up(db: Kysely<any>): Promise<void> {
     .addColumn('status', 'text', (col) => col.defaultTo('NOT_VERIFIED'))
     .addColumn('role', 'text', (col) => col.defaultTo('USER'))
     .addColumn('profile_picture_id', 'text')
-    .addColumn('verified_at', 'bigint')
-    .addColumn('created_at', 'bigint', (col) =>
+    .addColumn('verified_at', 'integer')
+    .addColumn('created_at', 'integer', (col) =>
       col.defaultTo(sql`(UNIXEPOCH())`).notNull()
     )
-    .addColumn('updated_at', 'bigint', (col) => col.defaultTo(null))
-    .addColumn('deleted_at', 'bigint', (col) => col.defaultTo(null))
+    .addColumn('updated_at', 'integer', (col) => col.defaultTo(null))
+    .addColumn('deleted_at', 'integer', (col) => col.defaultTo(null))
     .execute()
 }
